+++ lib/src/bones_api_entity_db_object_directory.dart
@@ -16,8 +16,9 @@ import 'bones_api_entity_db.dart';
 import 'bones_api_entity_db_object.dart';
 import 'bones_api_entity_reference.dart';
 import 'bones_api_extension.dart';
+import 'bones_api_logging.dart';
 
-final _log = logging.Logger('DBObjectDirectoryAdapter');
+final _log = logging.Logger('DBObjectDirectoryAdapter')..registerAsDbLogger();
 
 class DBObjectDirectoryAdapterContext
     implements Comparable<DBObjectDirectoryAdapterContext> {
@@ -48,6 +49,8 @@ class DBObjectDirectoryAdapter
   static void boot() {
     if (_boot) return;
     _boot = true;
+
+    DBObjectAdapter.boot();
 
     DBObjectAdapter.registerAdapter([
       'object.directory',