+++ lib/src/bones_api_entity_db_object_gcs.dart
@@ -21,8 +21,9 @@ import 'bones_api_entity_db.dart';
 import 'bones_api_entity_db_object.dart';
 import 'bones_api_entity_reference.dart';
 import 'bones_api_extension.dart';
+import 'bones_api_logging.dart';
 
-final _log = logging.Logger('DBObjectGCSAdapter');
+final _log = logging.Logger('DBObjectGCSAdapter')..registerAsDbLogger();
 
 class DBObjectGCSAdapterContext
     implements Comparable<DBObjectGCSAdapterContext> {
@@ -51,6 +52,8 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
   static void boot() {
     if (_boot) return;
     _boot = true;
+
+    DBObjectAdapter.boot();
 
     DBObjectAdapter.registerAdapter([
       'object.gcs',