+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 1.4.16
 
 - API Config:
-  - Allow  `log.all`, `log.error` and `log.db` to files. 
-- `DBSQLAdapter` implementations: `registerAsDbLogger`.
+  - Allow  `log.all`, `log.error` and `log.db` to files.
+- `DBAdapter`:
+  - Better `boot` hierarchy.
+  - `registerAsDbLogger` loggers of implementations of `DBAdapter`.
 - `LoggerHandler`:
   - Added `resolveLogDestiny` and `logBuffered`.
 - `DBAdapterException`: